+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/importpoc/ImportService.java
@@ -18,10 +18,11 @@
  ********************************************************************************/
 package org.eclipse.tractusx.traceability.assets.application.importpoc;
 
+import org.eclipse.tractusx.traceability.assets.domain.base.model.AssetBase;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
+import java.util.Map;
 
 public interface ImportService {
-    void importAssets(MultipartFile file);
+    Map<AssetBase, Boolean> importAssets(MultipartFile file);
 }